Wrap coordinates returned by getCenter
The coordinates returned by google maps' getCenter are not wrapped.
This means getCenter will return coordinates that are invalid from
Foursquare's point of view, causing Foursquare to return an error.

By using the result of getCenter to create a new LatLng object,
we get wrapped coordinates suitable for passing on to Foursquare.