Register knockout components in App
Before, ko.components.register was called within the components'
constructors.

This commit changes things so the component returns an object
representing the component. This object in turn is used as the argument
to ko.components.register.

One advantage of this is now the component can be tested separately
without the knockout library.