Replace all category objects with each data load
This commit removes the option 'key' mapping option directing the ko
mapping plugin to not replace category objects with each new set of data
from foursquare.

Having the key mapping option set causes the data mapping on everything
subsequent to the first mapping fail, so I'm removing the option for now
at least.

I had this in place because I'd like for a selected venue to remain
selected if it's still on the map when the map's dragend event fires.