Don't bounce markers when hovered over
Before, the markers bounced when hovered over on the map. This was not
good from a UI perspective; it was almost like the application was
playing a game of keep-away with the user when they moused over the
marker with the intention of clicking on it.

The trick to this is the list still needs to display a hovered state
when a map marker if hovered over.

The solution was to add a hoverOrigin member to the hoveredVenue
observable. Now, if a component subscribes to hoveredVenue, it can
check to see where the hover occurred and act accordingly.

In the case of the map component, it checks to see if the hover
happened on itself; if so, it doesn't animate the marker. If not, it
animates the marker safe in the knowledge that the user isn't even on
the map and no game of keep-away will take place.