Present preference dialog in run phase
Present the user with the option to use device geolocation in the run
phase of the application lifecycle as opposed to the init phase.

The run phase focusses on responding to events namely dragend on the
map and changes in the preference to use the device's geolocation
functionality.

It only makes sense to respond to events from the preference dialog in
the run phase as well.