Search for categories in addition to venues
Previously, the search was only looking at venue names. Now, if a
category name matches the search string, all the venues in that
category will match regardless of the venues' names matching.

This way, if a "coffee" is typed into the search field, all the venues
in the "Coffee Shops" category will show even if these venues don't
have "coffee" in their name (e.x., "Tim Hortons" and "Starbucks".)